feat(aya-obj): Support loading programs with ksyms
By default, BTF of extern functions has `extern` linkage. However,
verifier is not accepting functions with other linkages than `global`
and `static`.

Fixup of sizes and offsets of ksyms can't be retrieved from ELF.
Instead, the size of an unsigned integer is used for all types which
were found in the ksyms section.

Based on similar fixups which are done in libbpf.[0]

[0] https://github.com/torvalds/linux/commit/5bd022ec01f06

Fixes #1245