Prevent verification failure by making the log level static immutable
The log level implementation in #1335 was incomplete as the verifier
could reject programs which exceeded their instruction limits within
logging statements. This commit addresses this issue by making the
log level static variable immutable (s.t. the compiler puts it in a
read-only section) and adds an additional test which the verifier
will reject as an infinite loop iff it is unable to detect that the
static variable would otherwise allow the logging.