aya-ebpf: Remove `AyaBtfMapMarker`
It was never necessary in the first place. There is no place in the
kernel that enforces BTF map structs themselves to be anonymous. Only
pointer types (that are members of BTF maps) have to be anonymous, but
that's currently handled by LLVM[0].

BTF maps work just fine without it.

[0] https://github.com/llvm/llvm-project/pull/163174